WebApi/Controllers/UserController.cs
using Microsoft.AspNetCore.Mvc;
using FolderProjectApp.Components.HelperComponents;
using FolderProjectApp.Models.AuthModel;
using FolderProjectApp.Services.Interfaces;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;

namespace FolderProjectApp.Controllers
{
    [ApiController]
    [Route("[controller]")]
    public class UserController : ControllerBase
    {
        private readonly IUserAuthService _userService;

        public UserController(IUserAuthService userService)
        {
            _userService = userService;
        }

        [HttpPost("authenticate")]
        public async Task<IActionResult> Authenticate(AuthenticateRequest model)
        { 
            AuthenticateResponse response = await _userService.Authenticate(model);

            if (response == null)
                return BadRequest(new { message = "Username or password is incorrect " });

            return Ok(response);
        }

        [Authorize]
        [HttpGet]
        public async Task<IActionResult> GetAll()
        {
            var users = await _userService.GetAll();
            return  Ok(users);
        }
    }
}
